+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/test/java/org/apache/shardingsphere/proxy/backend/MockShardingSphereSchemasUtil.java
@@ -18,21 +18,14 @@
 package org.apache.shardingsphere.proxy.backend;
 
 import lombok.SneakyThrows;
-import org.apache.shardingsphere.infra.auth.Authentication;
-import org.apache.shardingsphere.infra.auth.ProxyUser;
-import org.apache.shardingsphere.kernal.context.SchemaContext;
-import org.apache.shardingsphere.kernal.context.SchemaContexts;
-import org.apache.shardingsphere.proxy.backend.schema.ProxySchemaContexts;
 import org.apache.shardingsphere.proxy.backend.schema.ShardingSphereSchema;
 import org.apache.shardingsphere.proxy.backend.schema.ShardingSphereSchemas;
 
 import java.lang.reflect.Field;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 public final class MockShardingSphereSchemasUtil {
     
@@ -47,24 +40,6 @@ public final class MockShardingSphereSchemasUtil {
         Field field = ShardingSphereSchemas.getInstance().getClass().getDeclaredField("schemas");
         field.setAccessible(true);
         field.set(ShardingSphereSchemas.getInstance(), mockSchemas(prefix, size));
-        Field schemaContexts = ProxySchemaContexts.getInstance().getClass().getDeclaredField("schemaContexts");
-        schemaContexts.setAccessible(true);
-        schemaContexts.set(ProxySchemaContexts.getInstance(), mockSchemaContexts(prefix, size));
-    }
-    
-    private static SchemaContexts mockSchemaContexts(final String prefix, final int size) {
-        SchemaContexts result = mock(SchemaContexts.class);
-        when(result.getAuthentication()).thenReturn(getAuthentication());
-        when(result.getSchemaContexts()).thenReturn(mockSchemaContextMap(prefix, size));
-        return result;
-    }
-    
-    private static Map<String, SchemaContext> mockSchemaContextMap(final String prefix, final int size) {
-        Map<String, SchemaContext> result = new HashMap<>(size);
-        for (int i = 0; i < size; i++) {
-            result.put(prefix + "_" + i, mock(SchemaContext.class));
-        }
-        return result;
     }
     
     private static Map<String, ShardingSphereSchema> mockSchemas(final String prefix, final int size) {
@@ -72,13 +47,6 @@ public final class MockShardingSphereSchemasUtil {
         for (int i = 0; i < size; i++) {
             result.put(prefix + "_" + i, mock(ShardingSphereSchema.class));
         }
-        return result;
-    }
-    
-    private static Authentication getAuthentication() {
-        ProxyUser proxyUser = new ProxyUser("root", Arrays.asList("schema_0", "schema_1"));
-        Authentication result = new Authentication();
-        result.getUsers().put("root", proxyUser);
         return result;
     }
 }
